service/src/main/java/uk/gov/dft/bluebadge/service/applicationmanagement/service/ApplicationService.java
package uk.gov.dft.bluebadge.service.applicationmanagement.service;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.util.Assert;
import uk.gov.dft.bluebadge.model.applicationmanagement.generated.Application;
import uk.gov.dft.bluebadge.service.applicationmanagement.converter.ApplicationConverter;
import uk.gov.dft.bluebadge.service.applicationmanagement.repository.ApplicationRepository;
import uk.gov.dft.bluebadge.service.applicationmanagement.repository.domain.ApplicationEntity;

@Slf4j
@Service
@Transactional
public class ApplicationService {

  private final ApplicationRepository repository;
  private final ApplicationConverter converter;

  @Autowired
  ApplicationService(ApplicationRepository repository, ApplicationConverter converter) {
    this.repository = repository;
    this.converter = converter;
  }

  public String createApplication(Application applicationModel) {
    ApplicationEntity application = converter.convertToEntityOnCreate(applicationModel);
    Assert.notNull(application.getId(), "Application Id must be populated before create.");
    int insertCount;
    log.info("Creating application: {}", application.getId());
    insertCount = repository.createApplication(application);
    log.debug("{} applications created", insertCount);

    repository.createHealthcareProfessionals(application.getHealthcareProfessionals());
    repository.createMedications(application.getMedications());
    repository.createTreatments(application.getTreatments());
    repository.createVehicles(application.getVehicles());
    repository.createWalkingAids(application.getWalkingAids());
    repository.createWalkingDifficultyTypes(application.getWalkingDifficultyTypes());
    return application.getId().toString();
  }
}
